separate utils dir; more ts
