+++ src/app/utils.ts
@@ -1,6 +1,4 @@
 import { fromSats } from 'satcomma';
-import { Event as NostrEvent, finishEvent, nip57, validateEvent, verifySignature } from 'nostr-tools';
-import { decode as invoiceDecode } from 'light-bolt11-decoder'
 
 export const validateLnurl = async (lnurl: string) => {
   const url = `/api/validate?lnurl=${encodeURIComponent(lnurl)}`;
@@ -30,144 +28,3 @@ export const handleStatusUpdate = (status, lnAddress, userAddress, jackpot, time
       toast(`Payment Failed - DM @_alexlewin on Twitter to get paid.`, { type: 'error', pauseOnFocusLoss: true });
     }
 };
-
-export const checkInvoiceStatus = (setChecking, hash, setHash, setSettled, toast, userAddress, setCountdownKey, newNote: NostrEvent | null, matchState) => {
-  setChecking(true);
-  let nostr: string | null = null
-  if (newNote && matchState.currentState === "WAITING") {
-    // tell server to post this specific "Round Started" kind 1
-    nostr = encodeURI(JSON.stringify(newNote))
-  }
-  const url = `/api/invoice?hash=${encodeURIComponent(hash!)}&lnaddr=${userAddress}&nostr=${nostr}`
-  console.debug('url', url)
-  fetch(url, { method: 'GET' })
-    .then((response) => response.json())
-    .then((data) => {
-      if (data.settled) {
-
-        setSettled(data.settled && true);
-        localStorage.setItem('lnaddr', userAddress);
-        setCountdownKey(prevKey => prevKey + 1);
-        setHash(null);
-        toast("Bid Received! You're in the lead!", { type: 'success' });
-      }
-      setChecking(false);
-    }).catch(_ => setChecking(false));
-}
-
-export async function getNewNostrPost(): Promise<{event: NostrEvent}> {
-  const url = '/api/nostr'
-  const res = await fetch(url, { method: 'POST' })
-  let data = await res.json()
-  console.debug('client getNewNostrPost data', data)
-  return {event: data}
-}
-
-export async function getZapEndpoint(
-  lud16: string | null,
-): Promise<null | { callback: string; minSendable: number; maxSendable: number; nostrPubkey: string; lnurl: string }> {
-  try {
-    let lnurl: string = ''
-    if (lud16) {
-      let [name, domain] = lud16.split('@')
-      lnurl = `https://${domain}/.well-known/lnurlp/${name}`
-    } else {
-      return null
-    }
-
-    let res = await fetch(lnurl)
-    let body = await res.json()
-
-    if (body.allowsNostr && body.nostrPubkey) {
-      return {
-        callback: body.callback,
-        minSendable: body.minSendable,
-        maxSendable: body.maxSendable,
-        nostrPubkey: body.nostrPubkey,
-        lnurl: lnurl,
-      }
-    }
-  } catch (err) {
-    /*-*/
-  }
-
-  return null
-}
-
-export const getZapInvoice = async (privateKey: string, nostrZapCallback: string, eventId: string): Promise<{invoice: string, paymentHash: string} | undefined> => {
-  const amountMillisats = parseInt(process.env.NEXT_PUBLIC_INVOICE_AMOUNT || "1000") * 1000
-
-  const zapRequestArgs = {
-    profile: process.env.NEXT_PUBLIC_NOSTR_HEX_PUBLIC_KEY!,
-    event: eventId,
-    amount: amountMillisats,
-    comment: 'Bid on lastpaywins.com',
-    // NOTE: lnbits lnurlp 0.3 will break if it tries to publish to non public relays I think
-    relays: ["wss://relay.damus.io"]
-  }
-  console.debug('zapRequestArgs', zapRequestArgs)
-
-  try {
-    const signedZapRequestEvent = finishEvent(nip57.makeZapRequest(zapRequestArgs), privateKey)
-    let ok = validateEvent(signedZapRequestEvent)
-    if (!ok) throw new Error('Invalid event')
-    let veryOk = verifySignature(signedZapRequestEvent)
-    if (!veryOk) throw new Error('Invalid signature')
-
-    const encodedZapRequest = encodeURI(JSON.stringify(signedZapRequestEvent))
-    const zapRequestHttp = `${nostrZapCallback}?amount=${amountMillisats.toString()}&nostr=${encodedZapRequest}&lnurl=${process.env.NEXT_PUBLIC_NOSTR_LIGHTNING_ADDRESS}`
-
-    const resObj = await fetch(zapRequestHttp).then((res) => res.json())
-    if (resObj.status === 'ERROR') throw new Error(resObj.reason)
-
-    const { pr: invoice } = resObj
-    const decodedInvoice = decodeInvoice(invoice)
-    if (!decodedInvoice?.paymentHash) throw new Error('Missing payment hash')
-
-    return { invoice, paymentHash: decodedInvoice.paymentHash } 
-  } catch(e) {
-    console.error(e)
-  }
-}
-
-export interface InvoiceDetails {
-  amount?: number
-  expire?: number
-  timestamp?: number
-  description?: string
-  descriptionHash?: string
-  paymentHash?: string
-  expired: boolean
-}
-
-const decodeInvoice = (pr: string): InvoiceDetails | undefined => {
-  try {
-    const parsed = invoiceDecode(pr)
-
-    const amountSection = parsed.sections.find((a) => a.name === 'amount')
-    const amount = amountSection ? Number(amountSection.value as number | string) : undefined
-    const timestampSection = parsed.sections.find((a) => a.name === 'timestamp')
-    const timestamp = timestampSection ? Number(timestampSection.value as number | string) : undefined
-    const expirySection = parsed.sections.find((a) => a.name === 'expiry')
-    const expire = expirySection ? Number(expirySection.value as number | string) : undefined
-    const descriptionSection = parsed.sections.find((a) => a.name === 'description')?.value
-    const descriptionHashSection = parsed.sections.find((a) => a.name === 'description_hash')?.value
-    const paymentHashSection = parsed.sections.find((a) => a.name === 'payment_hash')?.value
-
-    const ret = {
-      amount: amount,
-      expire: timestamp && expire ? timestamp + expire : undefined,
-      timestamp: timestamp,
-      description: descriptionSection as string | undefined,
-      descriptionHash: descriptionHashSection as string | undefined,
-      paymentHash: paymentHashSection as string | undefined,
-      expired: false,
-    }
-    if (ret.expire) {
-      ret.expired = ret.expire < new Date().getTime() / 1000
-    }
-    return ret
-  } catch (e) {
-    console.error(e)
-  }
-}
